re #1089 Adding headers to Xcode projects
Reducing warnings

+++ PlusLib/src/PlusCalibration/PatternLocAlgo/PlusFidLabeling.cxx
@@ -256,7 +256,7 @@ double PlusFidLabeling::ComputeSlope( PlusFidLine& line )
   double x = ( x2 - x1 );
 
   double t;
-  if ( fabsf( x ) > fabsf( y ) )
+  if ( fabs( x ) > fabs( y ) )
   {
     t = vtkMath::Pi() / 2 + atan( y / x );
   }
